deep refactor test
add 'testcontainers-annotations', remove 'easyway', migrate on JUnit 5
