--- src/test/java/com/rbkmoney/fistful/reporter/config/AbstractCephConfig.java
@@ -1,94 +1,0 @@
-package com.rbkmoney.fistful.reporter.config;
-
-import com.rbkmoney.easyway.*;
-import com.rbkmoney.fistful.reporter.config.properties.FileStorageProperties;
-import com.rbkmoney.fistful.reporter.config.properties.PartyManagementProperties;
-import com.rbkmoney.fistful.reporter.service.impl.FileStorageServiceImpl;
-import lombok.extern.slf4j.Slf4j;
-import org.junit.ClassRule;
-import org.junit.runner.Description;
-import org.junit.runner.RunWith;
-import org.springframework.boot.context.properties.EnableConfigurationProperties;
-import org.springframework.boot.test.context.ConfigFileApplicationContextInitializer;
-import org.springframework.boot.test.util.TestPropertyValues;
-import org.springframework.context.ConfigurableApplicationContext;
-import org.springframework.test.annotation.DirtiesContext;
-import org.springframework.test.context.ContextConfiguration;
-import org.springframework.test.context.TestPropertySource;
-import org.springframework.test.context.junit4.SpringRunner;
-import org.testcontainers.containers.FailureDetectingExternalResource;
-
-import java.util.function.Consumer;
-import java.util.function.Supplier;
-
-@RunWith(SpringRunner.class)
-@ContextConfiguration(
-        classes = {
-                ClientConfig.class,
-                FileStorageServiceImpl.class,
-        },
-        initializers = AbstractCephConfig.Initializer.class
-)
-@TestPropertySource("classpath:application.yml")
-@EnableConfigurationProperties(value = {FileStorageProperties.class, PartyManagementProperties.class})
-@DirtiesContext(classMode = DirtiesContext.ClassMode.AFTER_CLASS)
-@Slf4j
-public abstract class AbstractCephConfig extends AbstractTestUtils {
-
-    private static TestContainers testContainers = TestContainersBuilder
-            .builderWithTestContainers(getTestContainersParametersSupplier())
-            .addCephTestContainer()
-            .addFileStorageTestContainer()
-            .build();
-
-    @ClassRule
-    public static final FailureDetectingExternalResource resource = new FailureDetectingExternalResource() {
-
-        @Override
-        protected void starting(Description description) {
-            testContainers.startTestContainers();
-        }
-
-        @Override
-        protected void failed(Throwable e, Description description) {
-            log.warn("Test Container running was failed ", e);
-        }
-
-        @Override
-        protected void finished(Description description) {
-            testContainers.stopTestContainers();
-        }
-    };
-
-    public static class Initializer extends ConfigFileApplicationContextInitializer {
-
-        @Override
-        public void initialize(ConfigurableApplicationContext configurableApplicationContext) {
-            super.initialize(configurableApplicationContext);
-            TestPropertyValues.of(
-                    "kafka.topic.deposit.listener.enabled=false",
-                    "kafka.topic.destination.listener.enabled=false",
-                    "kafka.topic.source.listener.enabled=false",
-                    "kafka.topic.withdrawal.listener.enabled=false",
-                    "kafka.topic.wallet.listener.enabled=false",
-                    "kafka.topic.identity.listener.enabled=false")
-                    .and(testContainers.getEnvironmentProperties(getEnvironmentPropertiesConsumer()))
-                    .applyTo(configurableApplicationContext);
-        }
-    }
-
-    private static Supplier<TestContainersParameters> getTestContainersParametersSupplier() {
-        return () -> {
-            TestContainersParameters testContainersParameters = new TestContainersParameters();
-            testContainersParameters.setPostgresqlJdbcUrl("jdbc:postgresql://localhost:5432/fistful_reporter");
-
-            return testContainersParameters;
-        };
-    }
-
-    private static Consumer<EnvironmentProperties> getEnvironmentPropertiesConsumer() {
-        return environmentProperties -> {
-            environmentProperties.put("reporting.pollingEnable", "false");
-        };
-    }
-}
