+++ src/test/java/com/rbkmoney/fistful/reporter/config/initializer/ApplicationPropertiesInitializer.java
@@ -1,0 +1,14 @@
+package com.rbkmoney.fistful.reporter.config.initializer;
+
+import org.springframework.boot.test.context.ConfigDataApplicationContextInitializer;
+import org.springframework.boot.test.util.TestPropertyValues;
+import org.springframework.context.ConfigurableApplicationContext;
+
+public class ApplicationPropertiesInitializer extends ConfigDataApplicationContextInitializer {
+
+    @Override
+    public void initialize(ConfigurableApplicationContext configurableApplicationContext) {
+        TestPropertyValues.of("reporting.pollingEnable=false")
+                .applyTo(configurableApplicationContext);
+    }
+}
