+++ src/test/java/com/rbkmoney/fistful/reporter/config/testconfiguration/WithdrawalTestDaoConfig.java
@@ -1,0 +1,19 @@
+package com.rbkmoney.fistful.reporter.config.testconfiguration;
+
+import com.rbkmoney.fistful.reporter.dao.IdentityDao;
+import com.rbkmoney.fistful.reporter.dao.WalletDao;
+import com.rbkmoney.fistful.reporter.dao.WithdrawalDao;
+import org.springframework.boot.test.context.TestConfiguration;
+import org.springframework.context.annotation.Bean;
+
+@TestConfiguration
+public class WithdrawalTestDaoConfig {
+
+    @Bean
+    public WithdrawalTestDao withdrawalTestDao(
+            IdentityDao identityDao,
+            WalletDao walletDao,
+            WithdrawalDao withdrawalDao) {
+        return new WithdrawalTestDao(identityDao, walletDao, withdrawalDao);
+    }
+}
