+++ src/test/java/com/rbkmoney/fistful/reporter/dao/DepositDaoTest.java
@@ -1,0 +1,28 @@
+package com.rbkmoney.fistful.reporter.dao;
+
+import com.rbkmoney.fistful.reporter.config.PostgresqlSpringBootITest;
+import com.rbkmoney.fistful.reporter.domain.tables.pojos.Deposit;
+import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+
+import static com.rbkmoney.testcontainers.annotations.util.RandomBeans.random;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertNull;
+
+@PostgresqlSpringBootITest
+public class DepositDaoTest {
+
+    @Autowired
+    private DepositDao depositDao;
+
+    @Test
+    public void depositDaoTest() {
+        Deposit deposit = random(Deposit.class);
+        deposit.setCurrent(true);
+        Long id = depositDao.save(deposit).get();
+        deposit.setId(id);
+        assertEquals(deposit, depositDao.get(deposit.getDepositId()));
+        depositDao.updateNotCurrent(deposit.getId());
+        assertNull(depositDao.get(deposit.getDepositId()));
+    }
+}
