+++ src/test/java/com/rbkmoney/fistful/reporter/dao/IdentityDaoTest.java
@@ -1,0 +1,28 @@
+package com.rbkmoney.fistful.reporter.dao;
+
+import com.rbkmoney.fistful.reporter.config.PostgresqlSpringBootITest;
+import com.rbkmoney.fistful.reporter.domain.tables.pojos.Identity;
+import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+
+import static com.rbkmoney.testcontainers.annotations.util.RandomBeans.random;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertNull;
+
+@PostgresqlSpringBootITest
+public class IdentityDaoTest {
+
+    @Autowired
+    private IdentityDao identityDao;
+
+    @Test
+    public void identityDaoTest() {
+        Identity identity = random(Identity.class);
+        identity.setCurrent(true);
+        Long id = identityDao.save(identity).get();
+        identity.setId(id);
+        assertEquals(identity, identityDao.get(identity.getIdentityId()));
+        identityDao.updateNotCurrent(identity.getId());
+        assertNull(identityDao.get(identity.getIdentityId()));
+    }
+}
