+++ src/test/java/com/rbkmoney/fistful/reporter/handler/HandlerTest.java
@@ -2,28 +2,34 @@ package com.rbkmoney.fistful.reporter.handler;
 
 import com.rbkmoney.damsel.domain.Contract;
 import com.rbkmoney.fistful.reporter.*;
-import com.rbkmoney.fistful.reporter.config.AbstractHandlerConfig;
+import com.rbkmoney.fistful.reporter.config.PostgresqlSpringBootITest;
+import com.rbkmoney.fistful.reporter.config.testconfiguration.WithdrawalTestDao;
 import com.rbkmoney.fistful.reporter.generator.ReportGenerator;
 import com.rbkmoney.fistful.reporter.service.FileStorageService;
 import com.rbkmoney.fistful.reporter.service.PartyManagementService;
 import com.rbkmoney.fistful.reporter.service.ReportService;
 import org.apache.thrift.TException;
-import org.junit.Before;
-import org.junit.Test;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.mock.mockito.MockBean;
-import org.springframework.jdbc.core.JdbcTemplate;
 
 import java.io.IOException;
 import java.util.UUID;
 
+import static com.rbkmoney.fistful.reporter.data.TestData.contractId;
+import static com.rbkmoney.fistful.reporter.data.TestData.partyId;
 import static com.rbkmoney.geck.common.util.TypeUtil.temporalToString;
-import static org.junit.Assert.assertEquals;
+import static com.rbkmoney.testcontainers.annotations.util.ValuesGenerator.getFromTime;
+import static com.rbkmoney.testcontainers.annotations.util.ValuesGenerator.getToTime;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.Mockito.*;
 
-public class HandlerTest extends AbstractHandlerConfig {
+@PostgresqlSpringBootITest
+public class HandlerTest {
 
     @MockBean
     private PartyManagementService partyManagementService;
@@ -32,7 +38,7 @@ public class HandlerTest extends AbstractHandlerConfig {
     private FileStorageService fileStorageService;
 
     @Autowired
-    private JdbcTemplate jdbcTemplate;
+    private WithdrawalTestDao withdrawalTestDao;
 
     @Autowired
     private ReportService reportService;
@@ -45,7 +51,7 @@ public class HandlerTest extends AbstractHandlerConfig {
 
     private ReportRequest request;
 
-    @Before
+    @BeforeEach
     public void setUp() {
         ReportTimeRange reportTimeRange = new ReportTimeRange(
                 temporalToString(getFromTime()),
@@ -54,19 +60,19 @@ public class HandlerTest extends AbstractHandlerConfig {
         request = new ReportRequest(partyId, contractId, reportTimeRange);
     }
 
-    @Test(expected = InvalidRequest.class)
+    @Test
     public void exceptionArgTest() throws TException {
-        reportHandler.generateReport(request, "kek");
+        assertThrows(
+                InvalidRequest.class,
+                () -> reportHandler.generateReport(request, "kek"));
     }
 
     @Test
     public void fistfulReporterTest() throws TException, IOException {
-        jdbcTemplate.execute("truncate table fr.report cascade");
-
         when(partyManagementService.getContract(anyString(), anyString())).thenReturn(new Contract());
         when(fileStorageService.saveFile(any())).thenReturn(UUID.randomUUID().toString());
 
-        saveWithdrawalsDependencies();
+        withdrawalTestDao.saveWithdrawalsDependencies(5);
 
         long reportId = reportHandler.generateReport(request, "withdrawalRegistry");
 
@@ -77,11 +83,6 @@ public class HandlerTest extends AbstractHandlerConfig {
         assertEquals(ReportStatus.created, report.getStatus());
         assertEquals(1, report.getFileDataIds().size());
         verify(fileStorageService, times(1)).saveFile(any());
-    }
-
-    @Override
-    protected int getExpectedSize() {
-        return 5;
     }
 
     private void schedulerEmulation() {
