+++ src/test/java/com/rbkmoney/fistful/reporter/handler/SourceHandlerTest.java
@@ -1,23 +1,23 @@
 package com.rbkmoney.fistful.reporter.handler;
 
 import com.rbkmoney.fistful.account.Account;
-import com.rbkmoney.fistful.reporter.config.AbstractHandlerConfig;
+import com.rbkmoney.fistful.reporter.config.PostgresqlSpringBootITest;
 import com.rbkmoney.fistful.reporter.dao.SourceDao;
 import com.rbkmoney.fistful.reporter.dao.mapper.RecordRowMapper;
-import com.rbkmoney.fistful.reporter.domain.tables.pojos.Identity;
 import com.rbkmoney.fistful.reporter.domain.tables.pojos.Source;
 import com.rbkmoney.fistful.reporter.handler.source.SourceStatusChangedHandler;
-import org.junit.Test;
+import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.jdbc.core.JdbcTemplate;
 
 import static com.rbkmoney.fistful.reporter.domain.tables.Source.SOURCE;
-import static com.rbkmoney.fistful.reporter.utils.handler.SourceHandlerTestUtils.createMachineEvent;
-import static com.rbkmoney.fistful.reporter.utils.handler.SourceHandlerTestUtils.createStatusChanged;
-import static io.github.benas.randombeans.api.EnhancedRandom.random;
-import static org.junit.Assert.assertEquals;
+import static com.rbkmoney.fistful.reporter.util.handler.SourceHandlerTestUtil.createMachineEvent;
+import static com.rbkmoney.fistful.reporter.util.handler.SourceHandlerTestUtil.createStatusChanged;
+import static com.rbkmoney.testcontainers.annotations.util.RandomBeans.random;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 
-public class SourceHandlerTest extends AbstractHandlerConfig {
+@PostgresqlSpringBootITest
+public class SourceHandlerTest {
 
     @Autowired
     private SourceStatusChangedHandler sourceStatusChangedHandler;
@@ -32,7 +32,6 @@ public class SourceHandlerTest extends AbstractHandlerConfig {
     Account account = random(Account.class);
     String sqlStatement = "select * from fr.source where id='" + source.getId() + "';";
 
-
     @Test
     public void sourceAccountCreatedHandlerTest() {
         source.setCurrent(true);
@@ -45,11 +44,5 @@ public class SourceHandlerTest extends AbstractHandlerConfig {
                 jdbcTemplate.queryForObject(sqlStatement,
                         new RecordRowMapper<>(SOURCE, Source.class)).getCurrent()
         );
-    }
-
-
-    @Override
-    protected int getExpectedSize() {
-        return 0;
     }
 }
