+++ src/test/java/com/rbkmoney/fistful/reporter/handler/WithdrawalHandlerTest.java
@@ -1,20 +1,25 @@
 package com.rbkmoney.fistful.reporter.handler;
 
-import com.rbkmoney.fistful.reporter.config.AbstractHandlerConfig;
+import com.rbkmoney.fistful.reporter.config.PostgresqlSpringBootITest;
+import com.rbkmoney.fistful.reporter.dao.WithdrawalDao;
 import com.rbkmoney.fistful.reporter.dao.mapper.RecordRowMapper;
 import com.rbkmoney.fistful.reporter.domain.tables.pojos.Withdrawal;
-import com.rbkmoney.fistful.reporter.handler.withdrawal.*;
-import org.junit.Before;
-import org.junit.Test;
+import com.rbkmoney.fistful.reporter.handler.withdrawal.WithdrawalRouteChangeHandler;
+import com.rbkmoney.fistful.reporter.handler.withdrawal.WithdrawalStatusChangedHandler;
+import com.rbkmoney.fistful.reporter.handler.withdrawal.WithdrawalTransferCreatedHandler;
+import com.rbkmoney.fistful.reporter.handler.withdrawal.WithdrawalTransferStatusChangedHandler;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.jdbc.core.JdbcTemplate;
 
 import static com.rbkmoney.fistful.reporter.domain.tables.Withdrawal.WITHDRAWAL;
-import static com.rbkmoney.fistful.reporter.utils.handler.WithdrawalHandlerTestUtils.*;
-import static io.github.benas.randombeans.api.EnhancedRandom.random;
-import static org.junit.Assert.assertEquals;
+import static com.rbkmoney.fistful.reporter.util.handler.WithdrawalHandlerTestUtil.*;
+import static com.rbkmoney.testcontainers.annotations.util.RandomBeans.random;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 
-public class WithdrawalHandlerTest extends AbstractHandlerConfig {
+@PostgresqlSpringBootITest
+public class WithdrawalHandlerTest {
 
     @Autowired
     private WithdrawalStatusChangedHandler withdrawalStatusChangedHandler;
@@ -29,12 +34,15 @@ public class WithdrawalHandlerTest extends AbstractHandlerConfig {
     private WithdrawalRouteChangeHandler withdrawalRouteChangeHandler;
 
     @Autowired
+    private WithdrawalDao withdrawalDao;
+
+    @Autowired
     private JdbcTemplate jdbcTemplate;
 
     Withdrawal withdrawal = random(Withdrawal.class);
     String sqlStatement = "select * from fr.withdrawal where id='" + withdrawal.getId() + "';";
 
-    @Before
+    @BeforeEach
     public void setUp() {
         withdrawal.setCurrent(true);
         withdrawalDao.save(withdrawal);
@@ -86,11 +94,5 @@ public class WithdrawalHandlerTest extends AbstractHandlerConfig {
                 jdbcTemplate.queryForObject(
                         sqlStatement, new RecordRowMapper<>(WITHDRAWAL, Withdrawal.class)).getCurrent()
         );
-    }
-
-
-    @Override
-    protected int getExpectedSize() {
-        return 0;
     }
 }
