+++ src/test/java/com/rbkmoney/fistful/reporter/handler/destination/DestinationCreatedCryptoWalletHandlerTest.java
@@ -1,22 +1,22 @@
 package com.rbkmoney.fistful.reporter.handler.destination;
 
-import com.rbkmoney.fistful.reporter.config.AbstractHandlerConfig;
+import com.rbkmoney.fistful.reporter.config.PostgresqlSpringBootITest;
 import com.rbkmoney.fistful.reporter.dao.mapper.RecordRowMapper;
 import com.rbkmoney.fistful.reporter.domain.tables.pojos.Destination;
-import com.rbkmoney.fistful.reporter.utils.handler.DestinationHandlerTestUtils;
-import org.junit.Assert;
-import org.junit.Before;
-import org.junit.Test;
+import com.rbkmoney.fistful.reporter.util.handler.DestinationHandlerTestUtil;
+import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.jdbc.core.JdbcTemplate;
 
 import static com.rbkmoney.fistful.reporter.domain.tables.Destination.DESTINATION;
-import static com.rbkmoney.fistful.reporter.utils.handler.DestinationHandlerTestUtils.createCreated;
-import static com.rbkmoney.fistful.reporter.utils.handler.DestinationHandlerTestUtils.createCreatedMachineEvent;
-import static io.github.benas.randombeans.api.EnhancedRandom.random;
+import static com.rbkmoney.fistful.reporter.util.handler.DestinationHandlerTestUtil.createCreated;
+import static com.rbkmoney.fistful.reporter.util.handler.DestinationHandlerTestUtil.createCreatedMachineEvent;
+import static com.rbkmoney.testcontainers.annotations.util.RandomBeans.random;
 import static org.junit.Assert.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertNotNull;
 
-public class DestinationCreatedCryptoWalletHandlerTest extends AbstractHandlerConfig {
+@PostgresqlSpringBootITest
+public class DestinationCreatedCryptoWalletHandlerTest {
 
     @Autowired
     private DestinationCreatedHandler destinationCreatedHandler;
@@ -24,20 +24,15 @@ public class DestinationCreatedCryptoWalletHandlerTest extends AbstractHandlerCo
     @Autowired
     private JdbcTemplate jdbcTemplate;
 
-    Destination destination = random(Destination.class);
-    String sqlStatement = "select * from fr.destination LIMIT 1;";
-
-    @Before
-    public void setUp() {
-        destination.setCurrent(true);
-    }
-
     @Test
     public void destinationCreatedHandlerTest() {
+        Destination destination = random(Destination.class);
+        destination.setCurrent(true);
+        String sqlStatement = "select * from fr.destination LIMIT 1;";
         com.rbkmoney.fistful.base.Resource fistfulResource = new com.rbkmoney.fistful.base.Resource();
-        fistfulResource.setCryptoWallet(DestinationHandlerTestUtils.createResourceCryptoWallet());
+        fistfulResource.setCryptoWallet(DestinationHandlerTestUtil.createResourceCryptoWallet());
         com.rbkmoney.fistful.destination.Destination fistfulDestination
-                = DestinationHandlerTestUtils.createFistfulDestination(fistfulResource);
+                = DestinationHandlerTestUtil.createFistfulDestination(fistfulResource);
 
         destinationCreatedHandler.handle(
                 createCreated(fistfulDestination),
@@ -47,12 +42,7 @@ public class DestinationCreatedCryptoWalletHandlerTest extends AbstractHandlerCo
                 new RecordRowMapper<>(DESTINATION, Destination.class));
         assertEquals(true, destinationResult.getCurrent());
 
-        Assert.assertNotNull(destinationResult.getCryptoWalletId());
-        Assert.assertNotNull(destinationResult.getCryptoWalletCurrency());
-    }
-
-    @Override
-    protected int getExpectedSize() {
-        return 0;
+        assertNotNull(destinationResult.getCryptoWalletId());
+        assertNotNull(destinationResult.getCryptoWalletCurrency());
     }
 }
