+++ src/test/java/com/rbkmoney/fistful/reporter/kafka/listener/DepositEventListenerTest.java
@@ -6,34 +6,34 @@ import com.rbkmoney.fistful.deposit.StatusChange;
 import com.rbkmoney.fistful.deposit.TimestampedChange;
 import com.rbkmoney.fistful.deposit.status.Status;
 import com.rbkmoney.fistful.deposit.status.Succeeded;
-import com.rbkmoney.fistful.reporter.FistfulReporterApplication;
+import com.rbkmoney.fistful.reporter.config.KafkaPostgresqlSpringBootITest;
 import com.rbkmoney.fistful.reporter.dao.DepositDao;
 import com.rbkmoney.fistful.reporter.domain.enums.DepositStatus;
 import com.rbkmoney.fistful.reporter.domain.tables.pojos.Deposit;
 import com.rbkmoney.kafka.common.serialization.ThriftSerializer;
 import com.rbkmoney.machinegun.eventsink.SinkEvent;
-import lombok.extern.slf4j.Slf4j;
-import org.junit.Test;
-import org.junit.runner.RunWith;
+import com.rbkmoney.testcontainers.annotations.kafka.config.KafkaProducer;
+import org.apache.thrift.TBase;
+import org.junit.jupiter.api.Test;
 import org.mockito.ArgumentCaptor;
 import org.mockito.Captor;
-import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.test.mock.mockito.MockBean;
-import org.springframework.test.annotation.DirtiesContext;
-import org.springframework.test.context.junit4.SpringRunner;
 
+import static com.rbkmoney.fistful.reporter.data.TestData.machineEvent;
+import static com.rbkmoney.fistful.reporter.data.TestData.sinkEvent;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.Mockito.*;
 
-@Slf4j
-@DirtiesContext
-@RunWith(SpringRunner.class)
-@SpringBootTest(
-        classes = FistfulReporterApplication.class,
-        properties = {"kafka.state.cache.size=0"})
-public class DepositEventListenerTest extends AbstractListenerTest {
+@KafkaPostgresqlSpringBootITest
+public class DepositEventListenerTest {
 
-    private static final long MESSAGE_TIMEOUT = 10_000L;
+    @Value("${kafka.topic.deposit.name}")
+    private String topicName;
+
+    @Autowired
+    private KafkaProducer<TBase<?, ?>> testThriftKafkaProducer;
 
     @MockBean
     private DepositDao depositDao;
@@ -59,11 +59,10 @@ public class DepositEventListenerTest extends AbstractListenerTest {
                 .thenReturn(new Deposit());
 
         // When
-        produce(sinkEvent, "deposit");
-        Thread.sleep(MESSAGE_TIMEOUT);
+        testThriftKafkaProducer.send(topicName, sinkEvent);
 
         // Then
-        verify(depositDao, times(1))
+        verify(depositDao, timeout(5000).times(1))
                 .save(captor.capture());
         assertThat(captor.getValue().getDepositStatus())
                 .isEqualTo(DepositStatus.succeeded);
