+++ src/test/java/com/rbkmoney/fistful/reporter/kafka/listener/WithdrawalEventListenerTest.java
@@ -1,7 +1,7 @@
 package com.rbkmoney.fistful.reporter.kafka.listener;
 
 import com.rbkmoney.dao.DaoException;
-import com.rbkmoney.fistful.reporter.FistfulReporterApplication;
+import com.rbkmoney.fistful.reporter.config.KafkaPostgresqlSpringBootITest;
 import com.rbkmoney.fistful.reporter.dao.WithdrawalDao;
 import com.rbkmoney.fistful.reporter.domain.enums.WithdrawalStatus;
 import com.rbkmoney.fistful.reporter.domain.tables.pojos.Withdrawal;
@@ -12,28 +12,28 @@ import com.rbkmoney.fistful.withdrawal.status.Status;
 import com.rbkmoney.fistful.withdrawal.status.Succeeded;
 import com.rbkmoney.kafka.common.serialization.ThriftSerializer;
 import com.rbkmoney.machinegun.eventsink.SinkEvent;
-import lombok.extern.slf4j.Slf4j;
-import org.junit.Test;
-import org.junit.runner.RunWith;
+import com.rbkmoney.testcontainers.annotations.kafka.config.KafkaProducer;
+import org.apache.thrift.TBase;
+import org.junit.jupiter.api.Test;
 import org.mockito.ArgumentCaptor;
 import org.mockito.Captor;
-import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.test.mock.mockito.MockBean;
-import org.springframework.test.annotation.DirtiesContext;
-import org.springframework.test.context.junit4.SpringRunner;
 
+import static com.rbkmoney.fistful.reporter.data.TestData.machineEvent;
+import static com.rbkmoney.fistful.reporter.data.TestData.sinkEvent;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.Mockito.*;
 
-@Slf4j
-@DirtiesContext
-@RunWith(SpringRunner.class)
-@SpringBootTest(
-        classes = FistfulReporterApplication.class,
-        properties = {"kafka.state.cache.size=0"})
-public class WithdrawalEventListenerTest extends AbstractListenerTest {
+@KafkaPostgresqlSpringBootITest
+public class WithdrawalEventListenerTest {
 
-    private static final long MESSAGE_TIMEOUT = 10_000L;
+    @Value("${kafka.topic.withdrawal.name}")
+    private String topicName;
+
+    @Autowired
+    private KafkaProducer<TBase<?, ?>> testThriftKafkaProducer;
 
     @MockBean
     private WithdrawalDao withdrawalDao;
@@ -59,11 +59,10 @@ public class WithdrawalEventListenerTest extends AbstractListenerTest {
                 .thenReturn(new Withdrawal());
 
         // When
-        produce(sinkEvent, "withdrawal");
-        Thread.sleep(MESSAGE_TIMEOUT);
+        testThriftKafkaProducer.send(topicName, sinkEvent);
 
         // Then
-        verify(withdrawalDao, times(1))
+        verify(withdrawalDao, timeout(5000).times(1))
                 .save(captor.capture());
         assertThat(captor.getValue().getWithdrawalStatus())
                 .isEqualTo(WithdrawalStatus.succeeded);
