+++ src/test/java/com/rbkmoney/fistful/reporter/util/handler/SourceHandlerTestUtil.java
@@ -1,12 +1,12 @@
-package com.rbkmoney.fistful.reporter.utils.handler;
+package com.rbkmoney.fistful.reporter.util.handler;
 
 import com.rbkmoney.fistful.account.Account;
-import com.rbkmoney.fistful.wallet.*;
+import com.rbkmoney.fistful.source.*;
 import com.rbkmoney.kafka.common.serialization.ThriftSerializer;
 import com.rbkmoney.machinegun.eventsink.MachineEvent;
 import com.rbkmoney.machinegun.msgpack.Value;
 
-public class WalletHandlerTestUtils {
+public class SourceHandlerTestUtil {
 
     public static MachineEvent createMachineEvent(String id, Account account) {
         return new MachineEvent()
@@ -14,13 +14,12 @@ public class WalletHandlerTestUtils {
                 .setSourceId(id)
                 .setSourceNs("2")
                 .setCreatedAt("2021-05-31T06:12:27Z")
-                .setData(Value.bin(new ThriftSerializer<>().serialize("", createAccountCreated(account))));
+                .setData(Value.bin(new ThriftSerializer<>().serialize("", createStatusChanged(account))));
     }
 
-    public static TimestampedChange createAccountCreated(Account account) {
+    public static TimestampedChange createStatusChanged(Account account) {
         return new TimestampedChange()
                 .setOccuredAt("2021-05-31T06:12:27Z")
-                .setChange(Change.account(AccountChange.created(account)));
+                .setChange(Change.status(new StatusChange(Status.authorized(new Authorized()))));
     }
-
 }
