+++ src/test/java/com/rbkmoney/fistful/reporter/util/handler/WalletHandlerTestUtil.java
@@ -1,12 +1,14 @@
-package com.rbkmoney.fistful.reporter.utils.handler;
+package com.rbkmoney.fistful.reporter.util.handler;
 
 import com.rbkmoney.fistful.account.Account;
-import com.rbkmoney.fistful.source.*;
+import com.rbkmoney.fistful.wallet.AccountChange;
+import com.rbkmoney.fistful.wallet.Change;
+import com.rbkmoney.fistful.wallet.TimestampedChange;
 import com.rbkmoney.kafka.common.serialization.ThriftSerializer;
 import com.rbkmoney.machinegun.eventsink.MachineEvent;
 import com.rbkmoney.machinegun.msgpack.Value;
 
-public class SourceHandlerTestUtils {
+public class WalletHandlerTestUtil {
 
     public static MachineEvent createMachineEvent(String id, Account account) {
         return new MachineEvent()
@@ -14,13 +16,12 @@ public class SourceHandlerTestUtils {
                 .setSourceId(id)
                 .setSourceNs("2")
                 .setCreatedAt("2021-05-31T06:12:27Z")
-                .setData(Value.bin(new ThriftSerializer<>().serialize("", createStatusChanged(account))));
+                .setData(Value.bin(new ThriftSerializer<>().serialize("", createAccountCreated(account))));
     }
 
-    public static TimestampedChange createStatusChanged(Account account) {
+    public static TimestampedChange createAccountCreated(Account account) {
         return new TimestampedChange()
                 .setOccuredAt("2021-05-31T06:12:27Z")
-                .setChange(Change.status(new StatusChange(Status.authorized(new Authorized()))));
+                .setChange(Change.account(AccountChange.created(account)));
     }
-
 }
