+++ src/test/resources/META-INF/spring.factories
@@ -1,0 +1,3 @@
+# Application Context Initializers
+org.springframework.context.ApplicationContextInitializer=\
+com.rbkmoney.fistful.reporter.config.initializer.ApplicationPropertiesInitializer
