+++ src/test/resources/logback-test.xml
@@ -3,9 +3,20 @@
     <include resource="org/springframework/boot/logging/logback/defaults.xml"/>
     <include resource="org/springframework/boot/logging/logback/console-appender.xml"/>
 
-    <root level="warn">
+    <root level="info">
         <appender-ref ref="CONSOLE"/>
     </root>
-    <logger name="com.rbkmoney.woody" level="ALL"/>
-    <logger name="com.rbkmoney.fistful.reporter" level="ALL"/>
+    <logger name="com.rbkmoney.woody" level="ERROR"/>
+    <logger name="com.rbkmoney.fistful.reporter">
+        <level value="debug"/>
+    </logger>
+    <logger name="org.apache.kafka">
+        <level value="warn"/>
+    </logger>
+    <logger name="org.springframework.boot.test.context">
+        <level value="warn"/>
+    </logger>
+    <logger name="org.springframework.test.context">
+        <level value="warn"/>
+    </logger>
 </configuration>
